+++ src/ralph/api/__init__.py
@@ -1,4 +1,8 @@
 """Main module for Ralph's LRS API."""
+
+from contextlib import asynccontextmanager
+from inspect import iscoroutinefunction
+
 import sentry_sdk
 from fastapi import Depends, FastAPI
 
@@ -6,7 +10,7 @@ from ralph.conf import settings
 
 from .. import __version__
 from .auth import AuthenticatedUser, authenticated_user
-from .routers import health, statements
+from .routers import DATABASE_CLIENT, health, statements
 
 if settings.SENTRY_DSN is not None:
     sentry_sdk.init(
@@ -17,7 +21,26 @@ if settings.SENTRY_DSN is not None:
         max_breadcrumbs=50,
     )
 
-app = FastAPI()
+
+# pylint: disable=redefined-outer-name,unused-argument
+@asynccontextmanager
+async def lifespan(app: FastAPI):
+    """Handle application lifespan."""
+    # Startup actions goes here
+    # [...]
+
+    # We are now ready
+    yield
+
+    # Properly shutdown the application
+    if hasattr(DATABASE_CLIENT, "close") and callable(DATABASE_CLIENT.close):
+        # pylint: disable=expression-not-assigned
+        await DATABASE_CLIENT.close() if iscoroutinefunction(
+            DATABASE_CLIENT.close
+        ) else DATABASE_CLIENT.close()
+
+
+app = FastAPI(lifespan=lifespan)
 app.include_router(statements.router)
 app.include_router(health.router)