+++ src/ralph/api/routers/__init__.py
@@ -1,1 +1,11 @@
-# noqa: D104
+"""Routers related tools for the Ralph API."""
+
+from typing import Union
+
+from ralph.backends.database.base import BaseAsyncDatabase, BaseDatabase
+from ralph.conf import settings
+
+# pylint: disable=invalid-name
+DATABASE_CLIENT: Union[BaseDatabase, BaseAsyncDatabase] = getattr(
+    settings.BACKENDS.DATABASE, settings.RUNSERVER_BACKEND.upper()
+).get_instance()